Add UBI to at91bootstrap.
First pass of ubi integration.  The code all compiles.

It is not wired in, it doesn't get linked in.

Consider
$ diff -r driver/ubi/ ~/src/u-boot/drivers/mtd/ubi
$ diff -r driver/ubi/include/ ~/src/u-boot/include/

Second pass of ubi integration.  We can call ubi_init() now.

That means that we are actually linking in code from UBI now.

That meant adding in a handful more .c files, and a whole bunch
(probably all of the rest) of the .h files.

Started a ubi-glue.c that currently stubs out things we really need to
fill in sooner or later.

We needed to rename driver/ubi/common.h to driver/ubi/ubi-common.h because
at91bootstap already has a common.h, which conflicted.  (This involved
editing every file in the UBI source tree that included "common.h", and
probably mistouched files that included "blah-common.h".)

We limit the maximum number of UBI devices to 1, which means we don't
need to call sprintf("ubi%d", ...) in a couple of places.  (Which is
good, because we don't have sprintf().)

There were some conflicts in the prototypes for functions in <string.h>,
which were resolved in favor of not changing at91bootstrap.

Third pass of ubi integration.  ubi_init() now finds volumes.

This means that we are actually executing meaningful code in UBI that
actually reads the nandflash memory.

That meant adding in a couple of more files from the UBI code in u-boot
and tweaking some of the code that wasn't compiling nicely.

We added a new driver/ubiflash.c, parallel to driver/nandflash.c and
driver/spi_flash.c and friends, and modified load_kernel() to call
load_ubiflash() instead of load_nandflash().

In order to access the flash from the UBI code, we exposed functions
from nandflash.c.  This isn't terribly extensible to spi flash.  If we
ever decide to support UBI on SPI flash, we'll have to address this.

Fourth pass of ubi integration.  We can now boot the kernel.

...assuming that somebody a UBI image with a "kernel" volume and a "dtb"
volume in flash starting at address 0x20000 and going no further than
0x04000000.  (All of those parameters need to be KConfig-ified someday
soon.)

...and assuming that the kernel boot parameters string defined in the
at91bootstrap_defconfig file has been modified to point the kernel to
the proper MTD partition for the UBI image and the proper volume in that
UBI image.

...It's probably also a good idea to set ubi.fm_autoconvert=1 in the
kernel parameters so that the kernel enables fastmap for the image,
although I haven't determined if or how much that helps boot time.  It
should help though.

Fifth pass of ubi integration.  Add READONLY support.

Implement CONFIG_MTD_UBI_READONLY mode for UBI subsystem.

This elimated 13kB of code space from the image in my reference
implementation (even more with debug printf's disabled).

Sixth pass of ubi integration.  Wire in UBI support into the KConfig system.

Signed-off-by: Patrick Doyle <pdoyle@irobot.com>